Generate host programs with host_cc.
I realized it was silly to generate programs that run on the host build with the generated gcc, since in theory that will be a cross-compiler one day. Anyway, this nicely avoids the shared libs problem, even though I seem to have those working (though is still an issue, because the interpreter seems to require a full path, which doesn't make sense in the host context, I think).